requirements.txt
fastapi
uvicorn
requests
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-storage==2.19.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
openai>=1.50.0